Metric "db version" column turned to numeric for correctness. Fixes #26.
It worked also with "float" so no patching required, but just in case
for future.

+++ pgwatch2/sql/datastore_setup/config_store.sql
@@ -57,7 +57,7 @@ alter table pgwatch2.monitored_db add constraint preset_or_custom_config check
 create table pgwatch2.metric (
     m_id                serial primary key,
     m_name              text not null,
-    m_pg_version_from   float not null,
+    m_pg_version_from   numeric not null,
     m_sql               text not null,
     m_comment           text,
     m_is_active         boolean not null default 't',
